Refactor data types into folder
Why these changes are being introduced:

With a new data type "fulltexts" coming online soon, and the refactoring of
"records" and "embeddings" into equal footing data types, it felt like a good time
to establish a dedicated space for data type definitions.

How this addresses that need:

Adds new `data_types/` folder.  Records and Emmbeddings moved here, and Fulltexts
will live here soon.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-531
